Moves some mediafiles to media folder
