+++ client/src/index.scss
@@ -144,7 +144,7 @@ $theme-colors: map-merge($theme-colors, $custom-colors);
 }
 
 h1 {
-    font-size: 1em;
+    font-size: 3.25em;
     font-weight: bolder;
 }
